pyproject.toml
[tool.poetry]
name = "dcm-technical-task"
version = "0.1.0"
description = "Technical task for Python Backend Engineer posion"
authors = ["Eugen Ciur <eugen.ciur@ionos.com>", "Oleg Höfling <oleg.hoefling@ionos.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
celery = "^5.2.7"
djangorestframework = "^3.14.0"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pyfakefs = "^5.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
